requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-chart-js==0.0.3
django-chartjs==2.3.0
django-froala-editor==4.0.13
django-jazzmin==2.5.0
django-pandas==0.6.6
django-requests==0.0.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
urllib3==1.26.11
